+++ docs/release-notes.rst
@@ -1,5 +1,6 @@
 .. This work is licensed under a Creative Commons Attribution 4.0 International License.
 .. SPDX-License-Identifier: CC-BY-4.0
+.. Copyright (c) 2020 HCL Technologies Limited.
 .. Copyright (C) 2020 AT&T Intellectual Property
 
 Release Notes
@@ -9,6 +10,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on `Keep a Changelog <http://keepachangelog.com/>`__
 and this project adheres to `Semantic Versioning <http://semver.org/>`__.
+
+[0.0.5] - 2022-12-15
+--------------------
+* Database logic modification and improvements (`RICAPP-207 <https://jira.o-ran-sc.org/browse/RICAPP-207>`_)
 
 [0.0.4] - 2021-07-07
 --------------------